Fixed issue with input-connected DWConv features masker.
Made sure that the features of an input-connected DepthWise conv cannot
be optimized by PIT, since it would not make sense to completely discard
the information of some input channels. In fact, this should never
happen in normal PIT, but sometimes happens in PITSuperNet, when the DW
+ PW branch is not selected, which then causes issues during export.

Added corresponding unit test.

Lastly, removed some annoying warning messages in PIT about the fact
that dilation/rf are not optimized for Conv2D or Linear layers.